refactor(state): move finalized state test methods to a test module
